pyproject.toml
[project]
name = "aqi-prediction"
version = "0.1.0"
description = "AQI prediction using LSTM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "pandas>=2.1.0",
    "python-dotenv>=1.0.1",
    "torch>=2.1.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.2",
    "meteostat>=1.6.5",
    "beautifulsoup4>=4.12.3",
    "helium>=5.1.0",
    "watchfiles>=1.0.3",
    "pytest>=8.3.4",
]
